% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAnnotationFiles.R
\name{.get_cache}
\alias{.get_cache}
\title{wrapper function for getting BiocFileCache associated with uncoverapp package}
\usage{
.get_cache()
}
\value{
BiocFileCache object associated with uncoverappLib
}
\description{
wrapper function for getting BiocFileCache associated with uncoverapp package
}
