% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.uncoverapp.R
\name{uncoverAPP}
\alias{uncoverAPP}
\title{run.uncoverapp}
\usage{
uncoverAPP()
}
\value{
This return a Shiny App. The is no value
}
\description{
This function launches \code{unCOVERApp}, a
\code{Shiny} application for clinical assessment of sequence coverage.
Setting where uncoverapp will be launched with following \code{where} option:
`"browser`" in user default browser,  `"viewer`"  RStudio viewer and `"window`"
in a new Rstudio window.
}
\examples{
\dontrun{
file.name='../path/sorted.bed.gz'
tbi='.../path/sorted.bed.gz.tbi'
app()
}

## Only run this example in interactive R sessions

if (interactive()) {
app()
}


}
\author{
Emanuela Iovino
}
