% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAnnotationFiles.R
\name{getAnnotationFiles}
\alias{getAnnotationFiles}
\title{Get Annotation Files}
\usage{
getAnnotationFiles(assembly = c("hg19", "hg38"), verbose = FALSE)
}
\arguments{
\item{assembly}{(character) Reference genome: "hg19", "hg38"}

\item{verbose}{(logical) print messages}
}
\value{
Named list with paths to annotation files for each assembly
}
\description{
Downloads or retrieves annotation files for the specified genome assembly
}
\examples{
getAnnotationFiles(assembly = "hg38", verbose = TRUE)
}
