% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAnnotationFiles.R
\name{getAnnotationFiles}
\alias{getAnnotationFiles}
\title{download and rename sorted.bed.gz and sorted.bed.gz.tbi files for annotation of
low-coverage positions.}
\usage{
getAnnotationFiles(verbose = FALSE)
}
\arguments{
\item{verbose}{(logical) print messages}
}
\value{
(char) Path to local cached file
or initial download is required
}
\description{
download and rename sorted.bed.gz and sorted.bed.gz.tbi files for annotation of
low-coverage positions.
}
\examples{
getAnnotationFiles()
}
