% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.uncoverapp.R
\name{run.uncoverapp}
\alias{run.uncoverapp}
\title{Launch uncoverApp in specific environment}
\usage{
run.uncoverapp(where = c("browser", "viewer", "window"))
}
\arguments{
\item{where}{Character. One of \code{"browser"}, \code{"viewer"}, or \code{"window"}.
If missing, the default shiny launch behavior will be used.}
}
\value{
No value is returned. The function launches the Shiny app.
}
\description{
Controls the \code{shiny.launch.browser} option to launch uncoverApp
in an external web browser, the RStudio Viewer pane, or a new RStudio window.
}
\examples{
if (interactive()) {
  run.uncoverapp(where = "window")
}

}
