% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.uncoverapp.R
\name{run.uncoverapp}
\alias{run.uncoverapp}
\title{run.uncoverapp}
\usage{
run.uncoverapp()
}
\value{
This return `Shiny` App. The is no value
}
\description{
This function launches \code{unCOVERApp}, a \code{Shiny} application for clinical assessment of sequence coverage.
}
\examples{
#to run unCOVERApp use run.uncoverapp() function.

run.uncoverapp()

#After running `run.uncoverapp()`, users use *unCOVERApp*
}
\author{
Emanuela Iovino
}
\keyword{uncoverappLib}
