% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.uncoverapp.R
\name{run.uncoverapp}
\alias{run.uncoverapp}
\title{run.uncoverapp}
\usage{
run.uncoverapp()
}
\value{
This return a Shiny App. The is no value
}
\description{
This function launches \code{unCOVERApp}, a \code{Shiny} application for clinical assessment of sequence coverage.
}
\examples{
\dontrun{
run.uncoverapp()
to run unCOVERApp use run.uncoverapp() function.
}


#After running `run.uncoverapp()` the shiny app appears in your browser
}
\author{
Emanuela Iovino
}
