% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.uncoverapp.R
\name{run.uncoverapp}
\alias{run.uncoverapp}
\title{run.uncoverapp}
\usage{
run.uncoverapp(where = c("browser", "panel", "window"))
}
\arguments{
\item{where}{accept `"browser`" , `"panel`"  or `"window`". The option sets
where uncoverapp will be launched}
}
\value{
This return a Shiny App. The is no value
}
\description{
This function launches \code{unCOVERApp}, a
\code{Shiny} application for clinical assessment of sequence coverage.
Setting where uncoverapp will be launched with following \code{where} option:
`"browser`" in user default browser,  `"panel`"  RStudio viewer and `"window`"
in a new Rstudio window.
}
\examples{
\dontrun{
file.name='../path/sorted.bed.gz'
tbi='.../path/sorted.bed.gz.tbi'
run.uncoverapp(where="window")
}

## Only run this example in interactive R sessions

if (interactive()) {
run.uncoverapp(where="window")
}

#After running `run.uncoverapp(where="window")` the shiny app appears in your choosen location.


}
\author{
Emanuela Iovino
}
