% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.uncoverapp.R
\name{run.uncoverapp}
\alias{run.uncoverapp}
\title{Location for uncoverapp in RStudio enviroment}
\usage{
run.uncoverapp(where = c("browser", "viewer", "window"))
}
\arguments{
\item{where}{accept `"browser`" , `"viewer`"  or `"window`". The option sets
where uncoverapp will be launched. Using NULL, uncoverapp will use default
After running `run.uncoverapp(where="window")` the shiny app appears
in your chosen location.}
}
\value{
This return a  Shiny App. The is no value
}
\description{
This function controls the `shiny.launch.browser` option to launch uncoverapp
in an external `browser`, the RStudio viewer `"viewer"`, or a new `"window"` in
RStudio.
}
\examples{
## Only run this example in interactive R sessions

if (interactive()) {
run.uncoverapp(where="window")
}


}
