% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.uncoverapp.R
\name{run.uncoverapp}
\alias{run.uncoverapp}
\title{run.uncoverapp}
\usage{
run.uncoverapp()
}
\value{
This return a Shiny App. The is no value
}
\description{
This function launches \code{unCOVERApp}, a
\code{Shiny} application for clinical assessment of sequence coverage.
}
\examples{
\dontrun{
file.name='../path/sorted.bed.gz'
tbi='.../path/sorted.bed.gz.tbi'
run.uncoverapp()
}

## Only run this example in interactive R sessions

if (interactive()) {
run.uncoverapp()
}

#After running `run.uncoverapp()` the shiny app appears in your browser

}
\author{
Emanuela Iovino
}
