% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.uncoverapp.R
\name{uncoverAPP}
\alias{uncoverAPP}
\title{run.uncoverapp}
\usage{
uncoverAPP()
}
\value{
No value is returned. The function launches the Shiny app.
}
\description{
Launches \code{unCOVERApp}, a \code{Shiny} application for clinical assessment of sequence coverage.
The app can be launched in different environments using the \code{where} parameter:
\itemize{
  \item \code{"browser"} - launches in the user's default web browser,
  \item \code{"viewer"} - launches in the RStudio Viewer pane,
  \item \code{"window"} - launches in a new RStudio window.
}
}
\examples{
\dontrun{
  # Launch in RStudio viewer
  run.uncoverapp(where = "viewer")
  
  # Launch in browser
  run.uncoverapp(where = "browser")
}

## Only run this example in interactive R sessions
if (interactive()) {
  run.uncoverapp()
}

}
\author{
Emanuela Iovino
}
