% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncoverappLib-package.R
\docType{package}
\name{uncoverappLib-package}
\alias{uncoverappLib}
\alias{uncoverappLib-package}
\title{uncoverappLib: Interactive graphical application for clinical assessment
        of sequence coverage at the base-pair level}
\description{
a Shiny application containing a suite of graphical and statistical tools to
support clinical assessment of low coverage regions.It displays three web pages
each providing a different analysis module: Coverage analysis, calculate AF by
allele frequency app and binomial distribution.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Manuelaio/uncoverappLib}
  \item Report bugs at \url{https://github.com/Manuelaio/uncoverappLib/issues}
}
\seealso{
    Useful links:
        \itemize{
            \item \url{https://github.com/Manuelaio/uncoverappLib}
            \item Report bugs at
            \url{https://github.com/Manuelaio/uncoverappLib/issues}
        }

}
\author{
\strong{Maintainer}: Emanuela Iovino \email{emanuela.iovino@unibo.it}

Authors:
\itemize{
  \item Tommaso Pippucci \email{tommaso.pippucci@unibo.it}
}

}
\keyword{internal}
